refactor: decompose chat tab runtime ownership
