+++ src/features/chat/tabs/TabRuntimeFactory.ts
@@ -5,18 +5,34 @@ import type { Conversation } from '@/core/types';
 import type { TabAttention } from '@/features/chat/state/types';
 import type { FeatureHost } from '@/features/FeatureHost';
 
+import type {
+  PublishedTabRuntimeRef,
+  TabRuntimeCleanup,
+  TabRuntimeConstructionContext,
+  TabRuntimeControllerBundle,
+  TabRuntimeShellBundle,
+} from './runtime/TabRuntimeConstruction';
+import { buildTabRuntimeControllers } from './runtime/TabRuntimeControllers';
+import { buildTabRuntimeInputBindings } from './runtime/TabRuntimeInputBindings';
+import { buildTabRuntimeServices } from './runtime/TabRuntimeServices';
+import { buildTabRuntimeShell } from './runtime/TabRuntimeShell';
+import { buildTabRuntimeUI } from './runtime/TabRuntimeUI';
+import type { ForkContext } from './TabForking';
+import { registerTabRuntimeResourceOwner } from './TabLifecycle';
 import {
-  assembleTabRuntime,
-  type ForkContext,
-  type TabRuntimeCleanup,
-} from './Tab';
+  applyProviderUIGating,
+  refreshTabProviderUI,
+} from './TabProviderState';
 import type {
   AssembledTabRuntime,
   ProviderCatalogInfo,
   TabId,
+  TabInputBindings,
   TabProviderCatalogContext,
   TabRuntimeCleanupFailure,
   TabRuntimeResourceOwner,
+  TabServices,
+  TabUIComponents,
 } from './types';
 
 export interface TabRuntimeFactoryOptions {
@@ -107,6 +123,123 @@ export class TabRuntimeConstructionError extends Error {
     this.name = 'TabRuntimeConstructionError';
     this.rollbackFailures = rollbackFailures;
   }
+}
+
+function createPublishedTabRuntimeRef(): PublishedTabRuntimeRef {
+  let publishedRuntime: AssembledTabRuntime | null = null;
+  return {
+    requirePublished: () => {
+      if (!publishedRuntime) {
+        throw new Error('Tab runtime callback invoked before assembly completed');
+      }
+      return publishedRuntime;
+    },
+    current: () => publishedRuntime,
+    publish: (runtime) => {
+      if (publishedRuntime) {
+        throw new Error('Tab runtime was published more than once');
+      }
+      publishedRuntime = runtime;
+    },
+  };
+}
+
+function composeTabRuntime(
+  shell: TabRuntimeShellBundle,
+  services: TabServices,
+  ui: TabUIComponents,
+  controllerBundle: TabRuntimeControllerBundle,
+  inputBindings: TabInputBindings,
+  resourceOwner: TabRuntimeResourceOwner,
+): AssembledTabRuntime {
+  return {
+    session: shell.session,
+    get id() {
+      return shell.id;
+    },
+    get lifecycleState() {
+      return shell.lifecycleState;
+    },
+    set lifecycleState(value) {
+      shell.lifecycleState = value;
+    },
+    get hydrationState() {
+      return shell.hydrationState;
+    },
+    set hydrationState(value) {
+      shell.hydrationState = value;
+    },
+    get draftModel() {
+      return shell.draftModel;
+    },
+    set draftModel(value) {
+      shell.draftModel = value;
+    },
+    get providerId() {
+      return shell.providerId;
+    },
+    set providerId(value) {
+      shell.providerId = value;
+    },
+    get conversationId() {
+      return shell.conversationId;
+    },
+    set conversationId(value) {
+      shell.conversationId = value;
+    },
+    executionCoordinator: shell.executionCoordinator,
+    providerCatalogResolver: shell.providerCatalogResolver,
+    captureReviewableSettlement: shell.captureReviewableSettlement,
+    state: shell.state,
+    controllers: controllerBundle.controllers,
+    services,
+    ui,
+    dom: shell.dom,
+    renderer: controllerBundle.renderer,
+    inputBindings,
+    resources: {
+      get isDisposed() {
+        return resourceOwner.isDisposed;
+      },
+    },
+  };
+}
+
+function assembleTabRuntime(
+  options: TabRuntimeConstructionContext,
+): AssembledTabRuntime {
+  const runtimeRef = createPublishedTabRuntimeRef();
+  const shell = buildTabRuntimeShell(options, runtimeRef);
+  const services = buildTabRuntimeServices(shell, options, runtimeRef);
+  const ui = buildTabRuntimeUI(shell, services, options, runtimeRef);
+  const controllerBundle = buildTabRuntimeControllers(
+    shell,
+    services,
+    ui,
+    options,
+    runtimeRef,
+  );
+  const inputBindings = buildTabRuntimeInputBindings(
+    shell,
+    ui,
+    controllerBundle.controllers,
+    options,
+    runtimeRef,
+  );
+  const runtime = composeTabRuntime(
+    shell,
+    services,
+    ui,
+    controllerBundle,
+    inputBindings,
+    options.resourceOwner,
+  );
+  registerTabRuntimeResourceOwner(runtime, options.resourceOwner);
+  runtimeRef.publish(runtime);
+
+  refreshTabProviderUI(runtime, options.plugin);
+  applyProviderUIGating(runtime, options.plugin);
+  return runtime;
 }
 
 /** Creates a structurally complete tab runtime or rolls back every acquired resource. */
